Added: - "About" window - Language selection for notices and stuffs - Tons of documentation - Gold, SP, EXP, and STATS with success tracking - Auto character creation - Go to clientless after login into the game - Accepting and refusing Party - Party Matching - Server time & wheater - More data correctly tracked to Info class
Improved:
- Private messages to show
- Generated bot events
- Handling Client crash
- The minimap now can accept x,y,z and region coords
- SRAttributes trying to avoid duplicates
- Enum type will be used if we are sure about static values
- User information on Entity (Game Info) explorer
- PK2 Extractor

Fixed:
- Private message not sending
- Auto login from commands line

Changes:
- Coordinates will be handled as int

& more changes and bugfixes...

+++ xBot/Game/Database.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Collections.Specialized;
 using System.Data.SQLite;
 using System.IO;
@@ -117,6 +118,7 @@ namespace xBot.Game
 		{
 			if (db != null && db.State != System.Data.ConnectionState.Closed)
 			{
+				q.Dispose();
 				db.Close();
 				db = null;
       }
@@ -129,7 +131,10 @@ namespace xBot.Game
 		{
 			try
 			{
-				File.Delete("Data\\" + Name + ".db");
+				if (Exists(Name))
+				{
+					File.Delete("Data\\" + Name + ".db");
+				}
 				return true;
 			}
 			catch { return false; }
